Move transformUnits and transformSingle to IngameGameState from Game
