+++ agot-bg-game-server/src/common/ingame-game-state/action-game-state/resolve-march-order-game-state/combat-game-state/post-combat-game-state/after-winner-determination-game-state/renly-baratheon-ability-game-state/RenlyBaratheonAbilityGameState.ts
@@ -88,7 +88,7 @@ export default class RenlyBaratheonAbilityGameState extends GameState<
 
         selectedUnit.forEach(([region, footmenToRemove]) => {
             // Replace them by knight
-            const knightsToAdd = this.game.transformUnits(region, footmenToRemove, knight, this.entireGame);
+            const knightsToAdd = this.ingame.transformUnits(region, footmenToRemove, knight);
 
             if (houseCombatData.region == region) {
                 // In case the footman was party of the army,
