+++ agot-bg-game-server/src/common/ingame-game-state/westeros-game-state/wildlings-attack-game-state/crow-killers-nights-watch-victory-game-state/CrowKillersNightsWatchVictoryGameState.ts
@@ -62,7 +62,7 @@ export default class CrowKillersNightsWatchVictoryGameState extends GameState<Wi
 
     onSelectUnitsEnd(house: House, selectedUnits: [Region, Unit[]][]): void {
         selectedUnits.forEach(([region, footmen]) => {
-            this.game.transformUnits(region, footmen, knight, this.entireGame);
+            this.ingame.transformUnits(region, footmen, knight);
         });
 
         this.ingame.log({
